requirements.txt
Flask==3.0.3
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
Werkzeug==3.0.3